+++ CHANGELOG.md
@@ -9,8 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- `ArrDeque::clone` (Bit copy not clone)
-- `Drain::drop` (Double free) 
+- `ArrDeque::clone` (NG: bit copy, not clone)
+- `ArrDeque::drop` and `Drain::drop` (NG: double free) 
 
 ## [0.2.0] - 2026-06-18